// server/utils/interfaces/profile.interface.ts
export interface ProfileInterface {
  // userId: string
  firstName: string;
  lastName: string;
  phoneNumber: string;
  residentialAddress: string;
  officeAddress: string;
  city: string;
  state: string;
  country: string;
  areYouAnExistingCustomer: boolean;
  howDidYouHearAboutUs: string;
  modeOfIdentification: string;
  idDocument: string;
  passport: string;
  agentIdProof: string;
  currentOccupation: string;
}